+++ ResearchSuiteUI/ResearchSuiteUI/Info.plist
