+++ ResearchSuiteUI/ResearchSuiteUI/Localization/en.lproj/ResearchSuiteUI.strings
@@ -1,6 +1,6 @@
 /* 
-  ResearchSuite_UI.strings
-  ResearchSuite_UI
+  ResearchSuiteUI.strings
+  ResearchSuiteUI
 
   Copyright © 2016 Sage Bionetworks. All rights reserved.
