+++ ResearchSuiteUI/ResearchSuiteUI/Styleguide/UIColor+BridgeKeyNames.swift
@@ -1,6 +1,6 @@
 //
 //  UIColor+BridgeKeyNames.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
