+++ ResearchSuiteUI/ResearchSuiteUI/Utilities/StaticUtilities.swift
