+++ ResearchSuiteUI/ResearchSuiteUI/iOS/DebugStepViewController.swift
@@ -1,6 +1,6 @@
 //
 //  DebugStepViewController.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
