+++ ResearchSuiteUI/ResearchSuiteUI/iOS/Generic Step UI/RSDStepNavigationView.swift
@@ -1,6 +1,6 @@
 //
 //  RSDStepNavigationView.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Created by Josh Bruhin on 5/23/17.
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
