+++ ResearchSuiteUI/ResearchSuiteUI/iOS/RSDStepViewController.swift
@@ -1,6 +1,6 @@
 //
 //  RSDStepViewController.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
