+++ ResearchSuiteUI/ResearchSuiteUI/iOS/RSDTaskInfoViewController.swift
@@ -1,6 +1,6 @@
 //
 //  RSDTaskInfoViewController.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
