+++ ResearchSuiteUI/ResearchSuiteUI/iOS/Utilities/UIAlertController+Utilities.swift
@@ -1,6 +1,6 @@
 //
 //  UIAlertController+Utilities.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
