+++ ResearchSuiteUI/ResearchSuiteUI/iOS/Utilities/UIView+Layout.swift
@@ -1,6 +1,6 @@
 //
 //  UIView+LayoutExtensions.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
 //
