+++ ResearchSuiteUI/ResearchSuiteUI/iOS/Views/RSDDatePickerView.swift
@@ -1,6 +1,6 @@
 //
 //  RSDDatePickerView.swift
-//  ResearchSuite-UI
+//  ResearchSuiteUI
 //
 //  Created by Shannon Young on 10/23/17.
 //  Copyright © 2017 Sage Bionetworks. All rights reserved.
