+++ ResearchSuiteUI/ResearchSuiteUI/iOS/Views/RSDShadowGradient.swift
