+++ ResearchSuiteUI/ResearchSuiteUITests/Info.plist
